Track player total & handle zero-point rolls
Process totals during each player turn, ensuring accumulated score for
a turn is deducted should the player roll a zero-point roll.

> If a roll has zero points, then the player loses not only their turn, but
also accumulated score for that turn.

As scoring dice are removed from each successive scoring turn a player
has, the final roll must either score, such as getting a 1 or a 5, or be
left with no scoring dice to be able to retain the accumulated score for
that turn.

Having run a quick simulation, is shows this additional constraint makes
it much harder to get into the game.

```
Round	Player	Score	Total	Roll
----	------	-----	-----	----
0	mike	50	0	[2, 6, 5, 2, 4]
0	mike	50	0	[2, 4, 2, 5]
0	mike	50	0	[4, 6, 5]
0	mike	100	0	[6, 1]
0	mike	0	0	[6]                 # zero-point roll
0	bob	150	850	[5, 1, 3, 6, 3]     # reflects total
                                                    # achieved in round 1.
0	bob	100	850	[1, 3, 3]
0	bob	100	850	[1, 6]
0	bob	0	850	[2]                 # zero-point roll
1	mike	150	0	[4, 1, 5, 4, 3]
1	mike	0	0	[2, 6, 4]
1	bob	100	850	[6, 5, 5, 2, 4]
1	bob	600	850	[1, 5, 5, 5]
1	bob	50	850	[5, 6]
1	bob	100	850	[1]                 # scoring, hence bob
                                                    # gets to keep his
                                                    # score
1	bob	0	850	[]
```